Removal of bookmarking section
While bookmarking is of course useful during an OSINT investigation, it falls more under productivity tools, and not directly OSINT tools. This section was quite out of date, and I think that instead linking to a list like [jyguyomarch/awesome-productivity](https://github.com/jyguyomarch/awesome-productivity) at the end of the list, would be more valuable to readers.